[feat] : JWT 토큰 활용한 테스트를 위한 SwaggerConfig 관련 로직 추가
- apiKey : swagger OpenAPI에 보안 요소 추가
  JWT토큰을 인증에 활용해 API 테스트를 하기 위함
